+++ Assets/FunPlusSDK/Editor/XUPorter/XCodePostProcess.cs
@@ -14,7 +14,7 @@ public static class XCodePostProcess
 	public static void OnPostProcessBuild( BuildTarget target, string pathToBuiltProject )
 	{
 		if (target != BuildTarget.iOS) {
-			Debug.LogWarning("Target is not iPhone. XCodePostProcess will not run");
+			// Debug.LogWarning("Target is not iPhone. XCodePostProcess will not run");
 			return;
 		}