+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### v4.0.1-alpha.1
+
+* upgrade native Android SDK to v4.0.1-alpha.0
+* Remove the funsdk-default-config.json file.
+* Modify post process script to fit the new Android SDK.
+* Clean up some XUPorter warnings.
+
 ### v4.0.1-alpha.0
 
 * upgrade native iOS SDK to v4.0.3-alpha.0